Add 'uuid' dependency, create API endpoint for logs
In order to create unique ids for messages in logs, the 'uuid' library has been added as a dependency. Also, a new API endpoint for retrieving logs was created (/api/logs). This endpoint supports filtering logs by channel, tag, and limit parameters. On the other hand, a modification was also made in the .gitignore file to consider the logs directory. Finally, the package-lock.json file was updated after the addition of the new dependency.