Implement maximum log size in LogService
Added 'maxLogSize' parameter to the LogService to limit the length of added logs. This functionality has been implemented to prevent the system from crashing or slowing down due to extensively large logging messages. The 'message' will be trimmed if it exceeds the 'maxLogSize'. Also, the maximum log size was declared as a constant 'MAX_LOG_SIZE'.