Add debug query and improve Apollo client initialization
In this commit, the useDebugQuery is imported from generated graphql and used in the index.tsx file to fetch the debug data and it is displayed on the page. The Apollo client initialization is also enhanced. The getApolloClient function was updated to receive an optional cache parameter and default Apollo query options were added. Besides, a new useApolloClient hook was created to allow the Apollo client to be used as a hook in React. This hook is used in the _app.tsx file. The reason for these changes was to enhance the debug capability and optimize the use of the Apollo client in the application.