Remove page reload on logout
This commit removes the line of code which triggered a page reload upon user logout. This change was necessary to support the adoption of the newly implemented authentication store. Instead of reloading the whole page, we are now leveraging the reactivity of our auth store to update user status across the application. This significantly improves user experience by reducing unnecessary full-page reload and making logout action more responsive.