Refactor form submission to improve UX
Refactored auth pages(Login and Registration) to use form submission event instead of click event for handling user authentication. The form submit event provides a more semantic and accessible approach for handling forms compared to the button click. This way, we are able to leverage native form functionalities such as form validation and keyboard accessibility. The RightPanelWrapper has been refactored from a div to a form. This will allow user to submit form by hitting enter key and also ensures the form is submitted even when JavaScript is disabled, improving the overall accessibility and user experience.